Universalize Edge policy setters and extend for hardening Chrome

+++ ClearAllEdgePolicy.ps1
@@ -1,9 +1,19 @@
 function Clear-EdgePolicy {
-  # clear Edge policy settings by force deleting the appropriate registry keys
-  Remove-Item `
-    -Force `
-    -Recurse `
-    -Path 'HKLM:\SOFTWARE\Policies\Microsoft\Edge'
+  $PolicyPath = 'HKLM:\SOFTWARE\Policies\Microsoft\Edge'
+
+  # clear Edge policy settings by deleting the appropriate registry keys
+  if (Test-Path $PolicyPath 2> $null) {
+    
+    Remove-Item `
+      -Force `
+      -Recurse `
+      -Path $PolicyPath
+
+  } else {
+    Write-Warning `
+      "Key $($PolicyPath) does not exist! I've got nothing to clear. No changes were made."
+  }
+
 }
 
 Clear-EdgePolicy
